move runClass method to main
